requirements.txt
langchain
python-dotenv
requests
langchain-groq